Added level bounds, Created Bomb position tracking function
Static map now has bounds on every side. Created a function on the player manager class that registers the position of every bomb placed on the grid. This was a test performed in order to implement such function in the bots, this will be needed to create data for the fitness function.